@@ -7,33 +7,30 @@
     "@emotion/styled": "^11.10.6",
     "@material-ui/core": "^4.12.4",
     "@material-ui/icons": "^4.11.3",
-    "@material-ui/lab": "^4.0.0-alpha.61",
     "@mui/icons-material": "^5.11.11",
     "@mui/material": "^5.11.11",
-    "@mui/styles": "^5.13.2",
     "@testing-library/jest-dom": "^5.16.5",
-    "@testing-library/react": "^12.0.0",
+    "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
     "@vercel/analytics": "^0.1.11",
     "emailjs-com": "^3.2.0",
     "flowbite": "^1.6.3",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-github-calendar": "^3.4.0",
     "react-icons": "^4.8.0",
     "react-parallax-tilt": "^1.7.110",
     "react-pdf": "^6.2.2",
-    "react-reveal": "^1.2.0",
+    "react-reveal": "^1.2.2",
     "react-router-dom": "^6.8.2",
     "react-router-hash-link": "^2.4.3",
-    "react-scripts": "^5.0.1",
     "react-tsparticles": "^2.9.3",
     "reactjs-pdf-reader": "^1.0.12",
     "semantic-ui-css": "^2.5.0",
     "semantic-ui-react": "^2.1.4",
     "tsparticles": "^2.9.3",
     "typewriter-effect": "^2.19.0",
-    "validator": "^13.9.0",
+    "validator": "^13.11.0",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
@@ -42,7 +39,7 @@
   },
   "eslintConfig": {
     "extends": [
-      "react-app",
+      "react-app", 
       "react-app/jest"
     ]
   },
